Move package common -> hashes.common
