+++ hashes/common/helpers.py
